Update splashScreen branch
