+++ KuvarApp/SceneDelegate.swift
@@ -10,29 +10,19 @@ import UIKit
 class SceneDelegate: UIResponder, UIWindowSceneDelegate {
     
     var window: UIWindow?
-    
-    
-    func scene(_ scene: UIScene, willConnectTo session: UISceneSession, options connectionOptions: UIScene.ConnectionOptions) {
-        // Use this method to optionally configure and attach the UIWindow `window` to the provided UIWindowScene `scene`.
-        // If using a storyboard, the `window` property will automatically be initialized and attached to the scene.
-        // This delegate does not imply the connecting scene or session are new (see `application:configurationForConnectingSceneSession` instead).
-        
-        
-            guard let windowScene  = (scene as? UIWindowScene) else { return }
-            
-            let window = UIWindow(windowScene: windowScene)
-            
-            let navigationController = UINavigationController()
-            window.rootViewController = navigationController
-            window.makeKeyAndVisible()
-            self.window = window
-            
-            
-            
-            
-            
-        let mainCoordinator = MainCoordinator(window: window)
-            mainCoordinator.start()
+       var mainCoordinator: MainCoordinator?
+       
+       func scene(_ scene: UIScene, willConnectTo session: UISceneSession, options connectionOptions: UIScene.ConnectionOptions) {
+           guard let windowScene = (scene as? UIWindowScene) else { return }
+           
+           let window = UIWindow(windowScene: windowScene)
+           let navigationController = UINavigationController()
+           window.rootViewController = navigationController
+           window.makeKeyAndVisible()
+           self.window = window
+           
+           mainCoordinator = MainCoordinator(window: window)
+           mainCoordinator?.start()
         }
         
         func sceneDidDisconnect(_ scene: UIScene) {
